# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: CUPHY EduTech Platform API for Calcutta University Physics Honours Students
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /batches:
    get:
      operationId: listBatches
      tags: [batches]
      summary: List all batches
      parameters:
        - name: semester
          in: query
          schema:
            type: integer
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of batches
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchListResponse"
    post:
      operationId: createBatch
      tags: [batches]
      summary: Create new batch (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBatchBody"
      responses:
        "201":
          description: Created batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
  /batches/{batchId}:
    get:
      operationId: getBatchById
      tags: [batches]
      summary: Get batch by ID
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Batch detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchDetail"
    put:
      operationId: updateBatch
      tags: [batches]
      summary: Update batch (admin)
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBatchBody"
      responses:
        "200":
          description: Updated batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
    delete:
      operationId: deleteBatch
      tags: [batches]
      summary: Delete batch (admin)
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /batches/{batchId}/enroll:
    post:
      operationId: enrollInBatch
      tags: [batches]
      summary: Enroll user in batch (admin)
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollUserBody"
      responses:
        "200":
          description: Enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /batches/{batchId}/unenroll:
    post:
      operationId: unenrollFromBatch
      tags: [batches]
      summary: Remove user from batch (admin)
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollUserBody"
      responses:
        "200":
          description: Unenrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /lectures:
    get:
      operationId: listLectures
      tags: [lectures]
      summary: List lectures
      parameters:
        - name: subjectId
          in: query
          schema:
            type: integer
        - name: batchId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Lectures
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LectureListResponse"
    post:
      operationId: createLecture
      tags: [lectures]
      summary: Create lecture (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLectureBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lecture"
  /lectures/{lectureId}:
    get:
      operationId: getLectureById
      tags: [lectures]
      summary: Get lecture by ID
      parameters:
        - name: lectureId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lecture
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lecture"
    put:
      operationId: updateLecture
      tags: [lectures]
      summary: Update lecture (admin)
      parameters:
        - name: lectureId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLectureBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lecture"
    delete:
      operationId: deleteLecture
      tags: [lectures]
      summary: Delete lecture (admin)
      parameters:
        - name: lectureId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # NOTES
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List PDF notes
      parameters:
        - name: subjectId
          in: query
          schema:
            type: integer
        - name: batchId
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Notes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteListResponse"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create note (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{noteId}:
    put:
      operationId: updateNote
      tags: [notes]
      summary: Update note (admin)
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete note (admin)
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # TESTS
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Batch:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        semester:
          type: integer
        thumbnail:
          type: string
          nullable: true
        price:
          type: number
        originalPrice:
          type: number
          nullable: true
        paymentLink:
          type: string
          nullable: true
        isActive:
          type: boolean
        expiresAt:
          type: string
          nullable: true
        createdAt:
          type: string
        studentsCount:
          type: integer
      required: [id, name, semester, price, isActive, createdAt, studentsCount]
    BatchDetail:
      allOf:
        - $ref: "#/components/schemas/Batch"
        - type: object
          properties:
            subjects:
              type: array
              items:
                $ref: "#/components/schemas/Subject"
            totalLectures:
              type: integer
            totalNotes:
              type: integer
            totalTests:
              type: integer
    CreateBatchBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        semester:
          type: integer
        thumbnail:
          type: string
          nullable: true
        price:
          type: number
        originalPrice:
          type: number
          nullable: true
        paymentLink:
          type: string
          nullable: true
        isActive:
          type: boolean
        expiresAt:
          type: string
          nullable: true
      required: [name, semester, price, isActive]
    UpdateBatchBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        semester:
          type: integer
        thumbnail:
          type: string
          nullable: true
        price:
          type: number
        originalPrice:
          type: number
          nullable: true
        paymentLink:
          type: string
          nullable: true
        isActive:
          type: boolean
        expiresAt:
          type: string
          nullable: true
    EnrollUserBody:
      type: object
      properties:
        userId:
          type: integer
      required: [userId]
    BatchListResponse:
      type: object
      properties:
        batches:
          type: array
          items:
            $ref: "#/components/schemas/Batch"
        total:
          type: integer
      required: [batches, total]

    # SUBJECT SCHEMAS
    Subject:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        batchId:
          type: integer
        icon:
          type: string
          nullable: true
        order:
          type: integer
        lecturesCount:
          type: integer
        notesCount:
          type: integer
      required: [id, name, batchId, order, lecturesCount, notesCount]
...
    Lecture:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        youtubeVideoId:
          type: string
        subjectId:
          type: integer
        batchId:
          type: integer
        duration:
          type: string
          nullable: true
        order:
          type: integer
        isFree:
          type: boolean
        createdAt:
          type: string
      required: [id, title, youtubeVideoId, subjectId, batchId, order, isFree, createdAt]
    CreateLectureBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        youtubeVideoId:
          type: string
        subjectId:
          type: integer
        batchId:
          type: integer
        duration:
          type: string
          nullable: true
        order:
          type: integer
        isFree:
          type: boolean
      required: [title, youtubeVideoId, subjectId, batchId, order, isFree]
    LectureListResponse:
      type: object
      properties:
        lectures:
          type: array
          items:
            $ref: "#/components/schemas/Lecture"
      required: [lectures]

    # NOTE SCHEMAS
    Note:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        pdfUrl:
          type: string
        subjectId:
          type: integer
        batchId:
          type: integer
        category:
          type: string
          nullable: true
        isFree:
          type: boolean
        createdAt:
          type: string
      required: [id, title, pdfUrl, subjectId, batchId, isFree, createdAt]
    CreateNoteBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        pdfUrl:
          type: string
        subjectId:
          type: integer
        batchId:
          type: integer
        category:
          type: string
          nullable: true
        isFree:
          type: boolean
      required: [title, pdfUrl, subjectId, batchId, isFree]
    NoteListResponse:
      type: object
      properties:
        notes:
          type: array
          items:
            $ref: "#/components/schemas/Note"
      required: [notes]

    # TEST SCHEMAS
...