# lib/api-zod/src/generated/types/lecture.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CUPHY EduTech Platform API for Calcutta University Physics Honours Students
 * OpenAPI spec version: 1.0.0
 */

export interface Lecture {
  id: number;
  title: string;
  description?: string | null;
  youtubeVideoId: string;
  subjectId: number;
  batchId: number;
  duration?: string | null;
  order: number;
  isFree: boolean;
  createdAt: string;
}