# lib/api-zod/src/generated/types/updateBatchBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CUPHY EduTech Platform API for Calcutta University Physics Honours Students
 * OpenAPI spec version: 1.0.0
 */

export interface UpdateBatchBody {
  name?: string;
  description?: string;
  semester?: number;
  thumbnail?: string | null;
  price?: number;
  originalPrice?: number | null;
  paymentLink?: string | null;
  isActive?: boolean;
  expiresAt?: string | null;
}